Move MetricLogger from fdbclient to fdbserver
